Remove conversion warning

+++ include/definitions.h
@@ -48,7 +48,7 @@ struct game {
     bool ended;
     int winner; // Winner's player index,-1 if no winner
     struct player players[MAX_PLAYERS];
-    int player_count;
+    long player_count;
     char map[MAP_WIDTH][MAP_HEIGHT];
 };
